// src/shared/components/ShadcnSelect.d.ts
import type { ComponentType } from "react";

declare const ShadcnSelect: ComponentType<{
  value?: any;
  onValueChange?: any;
  options?: any[];
  placeholder?: string;
  ariaLabel?: any;
  disabled?: boolean;
  className?: any;
  triggerClassName?: any;
  contentClassName?: any;
  itemClassName?: any;
  name?: any;
  [key: string]: any;
}>;

export default ShadcnSelect;